+++ Src/DiKErnel.Integration/Data/GrassRevetmentOvertopping/GrassRevetmentOvertoppingLocationDependentInput.cs
@@ -154,8 +154,9 @@ namespace DiKErnel.Integration.Data.GrassRevetmentOvertopping
 
             if (verticalDistanceWaterLevelElevation >= 0.0)
             {
-                int incrementTime = RevetmentFunctions.IncrementTime(timeDependentInput.BeginTime,
-                                                                     timeDependentInput.EndTime);
+                double incrementTime = RevetmentFunctions.IncrementTime(timeDependentInput.BeginTime,
+                                                                        timeDependentInput.EndTime);
+
                 averageNumberOfWaves = RevetmentFunctions.AverageNumberOfWaves(incrementTime,
                                                                                timeDependentInput.WavePeriodTm10,
                                                                                AverageNumberOfWavesCtm);
